- Introduced the use of gethostbyname_r() if available. As siproxd   uses threads it probably is a very good idea (some funny things   seen with Linux 2.4.18 kernel & RTP proxy - gethostbyname()   did return a NULL result but has h_errno set to 0 "every fine, thanks")   Future will show it this helps. - Had overlooked one inet_aton in register.c - replaced to utils_inet_aton - Fixed a compiler warning in log.c for Solaris. Siproxd now at least   builds on Solaris (tested on: SunOS 5.9 sun4u sparc SUNW,Ultra-60)

+++ src/log.c
@@ -205,7 +205,7 @@ void log_dump_buffer(int class, char *file, int line,
       for (j=0;(j<16) && (i+j)<length ;j++) {
          sprintf(tmp,"%2.2x ",(unsigned char)buffer[i+j]);
          strcat(tmplin1, tmp);
-         sprintf(tmp, "%c",(isprint(buffer[i+j]))? buffer[i+j]: '.');
+         sprintf(tmp, "%c",(isprint((int)buffer[i+j]))? buffer[i+j]: '.');
          strcat(tmplin2, tmp);
       }
       fprintf(stderr, "  %-47.47s %-16.16s\n",tmplin1, tmplin2);
